requirements.txt
numpy
scipy
colour-science
matplotlib